Simplifies fpswhen by using an auxilliary node
The node has the switch and the output node as parents and sets up the next tick (and/or cancels the timer that pushes to the output node) whenever either node is pushed to.

This enables simplifying the test for whether a node's action is required to a just checking for whether the node has actions, and if any parents were active.